Move Route class to Support\Models namespace
While I think the namespace Hyde\Routing\Route is beautiful I don't think it's needed for just a few classes, and I don't think Hyde will get more Routing classes in the foreseeable future.
