+++ packages/framework/src/Foundation/RouteCollection.php
@@ -6,7 +6,7 @@ namespace Hyde\Foundation;
 
 use Hyde\Foundation\Concerns\BaseFoundationCollection;
 use Hyde\Framework\Concerns\HydePage;
-use Hyde\Routing\Route;
+use Hyde\Support\Models\Route;
 
 /**
  * Pseudo-Router for Hyde.
